# Improve browscap get_browser performance