--- a/UPGRADING
+++ b/UPGRADING
@@ -197,3 +197,5 @@ PHP 8.4 UPGRADE NOTES
   They now run in linear time instead of being bounded by quadratic time.
 
 * mb_strcut() is much faster now for UTF-8 and UTF-16 strings.
+
+* get_browser() is much faster now, up to 1.5x - 2.5x for some test cases.